Break syndetics helper methods into separate module
